Cleaned up /dialogs/ file structure.
