+++ packages/app/client/src/ui/dialogs/index.ts
@@ -1,0 +1,5 @@
+export * from './botCreationDialog';
+export * from './host';
+export * from './secretPromptDialog';
+export * from './tabManager';
+export * from './service';
