+++ packages/app/client/src/ui/shell/main.tsx
@@ -39,10 +39,9 @@ import { Colors, Fonts, Splitter } from '@bfemulator/ui-react';
 import ExplorerBar from './explorer';
 import MDI from './mdi';
 import NavBar from './navBar';
-import TabManager from '../dialogs/tabManager';
+import { TabManager, DialogHost } from '../dialogs';
 import * as Constants from '../../constants';
 import StatusBar from './statusBar';
-import DialogHost from '../dialogs/host';
 import StoreVisualizer from '../debug/storeVisualizer';
 import store from '../../data/store';
 import { KeyCodes } from '@uifabric/utilities/lib';
